Add: Moderate overhaul of install script
Moved creation of the temp directory before the formatting of the
storage since formatting should not occur if temp directory cannot be
made.

Added ability to select between 32 bit (armv7l) and 64 bit (AArch64)
versions with default to 32 bit and instructions for making the
selection.

Added instructions to the end of the script. If installing 64 bit, it
also displays a message about the known USB issue on the 8 GB Pi 4 and
the workaround.

Installer now changes to the temp directory once it is set up. Still
uses absolute paths for everything, but the change is done to enable
md5sum.

Installer now downloads the .md5 of the tarball and verifies before
installing.